Allow multiple categories
Even though Steam's interface doesn't allow multiple categories for a
single game, it does have two parallel categories system: the user's and
the Favorites. Also, this part is called "tags" in the config file, and
the values are numbered, so it's better if we just support that out of
the box.