Change backup system to include the clean image's hash
The old backup system was to simply copy the "clean" (no overlays) file
to "{id} (original).jpg". When reapplying overlays, this clean file is
used as base. This works perfectly, but only when SteamGrid is the only
program touching these files.

If the user customizes an image directly on Steam (right click -> set
custom image), Steam changes the "{id}.jpg" file. But SteamGrid doesn't
know the file changed, and promptly deletes and reloads from the
"(original)" backup, discarding the user's change. This bug has wasted a
lot of users' time and effort. Sorry about that :(

The solution is to name the backup file (previously "{id}
(original).jpg") to include the overlaid file hash (now "{id} backup
{sha256}.jpg"). Now when processing a file, first SteamGrid reads
"{id}.jpg" and takes it hash. If there is a backup file with this hash,
it means the file was created by SteamGrid, and the backup file is the
just the clean version. If the user manually customizes an image, the
hash won't match any backups, and SteamGrid won't overwrite it.

The downside is that it may create files that will never be cleaned. As
the images are small, this shouldn't be an issue. Maybe a newer version
will automatically cleanup unused backups.

This commit also looks for old-style, legacy backups ("{id}
(original).jpg"), and converts them to new-style backups.