Completion of non-listed commands, and little typo check.
I’ve added most not ¨documented¨ (On the README.md atleast) launch commands, that i feel, is useful for users who want to customize even more steamgrid usage. That’s why I haven’t the `-steamdir` one, which is more of a troubleshooting one, in case it doesn’t find it. I could add it as an alternative method to the troubleshooting `Fails to find steam location` entry to complete it even further.
Also took the time to check for typos and little corrections to do. Some are extremely pointless and futile, as correctly writing the macOS according to their stylized name, adding a couple of commas, correcting apostrophes and correctly writing size units per regulation.