Handle WEBP animations.
Requires use of a package github.com/kmicki/webpanimation which is
forked from github.com/size-of-int/webpanimation. The fork adds
necessary functions for decoding WEBP animation into single frames
(original package only supports encoding single frames into an
animation).

The webpanimation package uses cgo to bind to libwebp, so gcc is
necessary to build the application from now on.

It is working on Linux (tested on Arch) and Windows (tested on Windows
10 with mingw installed). It was not tested on a Mac but should
also work.

Allows use of webp content. The extension is changed to .png when
the file is saved anyway, otherwise Steam will not use it.
But it works with a changed extension.